+++ src/lib/pagewise-chat-transport.ts
@@ -10,7 +10,7 @@ import {
 import { resolveStreamingTransform } from "./stream-transform";
 import { clearAgentProgress, subscribeAgentProgress } from "./agent-progress";
 import { wrapStreamWithAgentProgress } from "./inject-progress-stream";
-import { setAgentRunAbortSignal, clearAgentRunAbortSignal } from "./vision-index";
+import { setAgentRunAbortSignal } from "./vision-index";
 import {
   createUsageMetadataTracker,
   type PageWiseMessageMetadata,
@@ -95,7 +95,6 @@ export class PagewiseChatTransport<
       return wrapStreamWithAgentProgress(uiStream, earlyProgress);
     } finally {
       unsubEarly();
-      clearAgentRunAbortSignal();
     }
   }